+++ sdk/monitor/Azure.Monitor.OpenTelemetry.AspNetCore/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features Added
 
-* Logs emitted within an Activity context will now be sampled by default based on the `SamplingRatio` configured for Traces. To disable this sampling set `DisableTraceBasedSamplingForLogs` to `true`.
+* Logs emitted within an Activity context will now be sampled by default based on the `SamplingRatio` configured for Traces. To disable this sampling set `EnableTraceBasedSamplingForLogs` to `false`.
   ([#43793](https://github.com/Azure/azure-sdk-for-net/pull/43793))
 
 ### Breaking Changes